refactor(memory_backup): promote to directory, split tests (#1266)
- src/memory_backup.rs: 730 → src/memory_backup/mod.rs (384 LOC)

Pure refactor: 11 tests pass.
